- removed @Data annotation, @Data annotation not working

+++ src/main/java/examples/demo/repository/ReceiptRepo.java
@@ -2,14 +2,14 @@ package examples.demo.repository;
 
 
 
-import examples.demo.DatabaseModel.ReceiptDatabaseModel;
+import examples.demo.databaseModel.ReceiptDatabaseModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
 
 @Repository
-public interface ReceiptRepo extends JpaRepository<ReceiptDatabaseModel, Integer> {
+public interface ReceiptRepo extends JpaRepository<ReceiptDatabaseModel, Long> {
 
 
 }
